# Rtree formula bench