--- a/go.mod
+++ b/go.mod
@@ -7,7 +7,7 @@ require (
 	github.com/apache/arrow/go/arrow v0.0.0-20200403134915-89ce1cadb678
 	github.com/astrogo/fitsio v0.1.0
 	github.com/campoy/embedmd v1.0.0
-	github.com/containous/yaegi v0.8.2
+	github.com/cosmos72/gomacro v0.0.0-20200226185326-30b9859a00de
 	github.com/golang/freetype v0.0.0-20170609003504-e2365dfdc4a0
 	github.com/gonuts/binary v0.2.0
 	github.com/gonuts/commander v0.1.0
@@ -18,6 +18,7 @@ require (
 	github.com/jcmturner/gokrb5/v8 v8.3.0
 	github.com/jung-kurt/gofpdf v1.16.2 // indirect
 	github.com/klauspost/compress v1.10.3
+	github.com/mattn/go-runewidth v0.0.9 // indirect
 	github.com/peterh/liner v1.2.0
 	github.com/pierrec/lz4 v2.4.1+incompatible
 	github.com/pierrec/xxHash v0.1.5
@@ -29,7 +30,8 @@ require (
 	golang.org/x/exp v0.0.0-20200228211341-fcea875c7e85
 	golang.org/x/image v0.0.0-20200119044424-58c23975cae1 // indirect
 	golang.org/x/mobile v0.0.0-20200222142934-3c8601c510d0
-	golang.org/x/sync v0.0.0-20190423024810-112230192c58
+	golang.org/x/sync v0.0.0-20190911185100-cd5d95a43a6e
+	golang.org/x/tools v0.0.0-20200502202811-ed308ab3e770 // indirect
 	gonum.org/v1/gonum v0.7.1-0.20200330111830-e98ce15ff236
 	gonum.org/v1/plot v0.7.1-0.20200414075901-f4e1939a9e7a
 	gopkg.in/yaml.v3 v3.0.0-20200121175148-a6ecf24a6d71