This fixes the build and provides interesting performance improvements:

```
$> benchstat old.txt new.txt
name                           old time/op    new time/op    delta
FormulaEval/42.0-4                776ns ± 4%     310ns ± 3%  -60.05%  (p=0.000 n=20+20)
FormulaEval/F64-4                1.06µs ± 4%    0.34µs ± 2%  -68.15%  (p=0.000 n=19+17)
FormulaEval/2*F64-4              1.11µs ± 2%    0.34µs ± 1%  -69.83%  (p=0.000 n=18+19)
FormulaEval/math.Abs(2*F64)-4    1.73µs ± 2%    0.35µs ± 2%  -79.79%  (p=0.000 n=18+19)
FormulaFunc/42.0-4                564ns ± 3%     322ns ± 4%  -42.89%  (p=0.000 n=18+20)
FormulaFunc/F64-4                 834ns ± 3%     342ns ± 1%  -58.95%  (p=0.000 n=18+16)
FormulaFunc/2*F64-4               935ns ± 3%     344ns ± 2%  -63.18%  (p=0.000 n=17+20)
FormulaFunc/math.Abs(2*F64)-4    1.45µs ± 2%    0.36µs ± 1%  -75.05%  (p=0.000 n=17+18)

name                           old alloc/op   new alloc/op   delta
FormulaEval/42.0-4                 280B ± 0%       72B ± 0%  -74.29%  (p=0.000 n=20+20)
FormulaEval/F64-4                  384B ± 0%       72B ± 0%  -81.25%  (p=0.000 n=20+20)
FormulaEval/2*F64-4                408B ± 0%       72B ± 0%  -82.35%  (p=0.000 n=20+20)
FormulaEval/math.Abs(2*F64)-4      552B ± 0%       72B ± 0%  -86.96%  (p=0.000 n=20+20)
FormulaFunc/42.0-4                 240B ± 0%       64B ± 0%  -73.33%  (p=0.000 n=20+20)
FormulaFunc/F64-4                  344B ± 0%       64B ± 0%  -81.40%  (p=0.000 n=20+20)
FormulaFunc/2*F64-4                368B ± 0%       64B ± 0%  -82.61%  (p=0.000 n=20+20)
FormulaFunc/math.Abs(2*F64)-4      512B ± 0%       64B ± 0%  -87.50%  (p=0.000 n=20+20)

name                           old allocs/op  new allocs/op  delta
FormulaEval/42.0-4                 8.00 ± 0%      3.00 ± 0%  -62.50%  (p=0.000 n=20+20)
FormulaEval/F64-4                  11.0 ± 0%       3.0 ± 0%  -72.73%  (p=0.000 n=20+20)
FormulaEval/2*F64-4                12.0 ± 0%       3.0 ± 0%  -75.00%  (p=0.000 n=20+20)
FormulaEval/math.Abs(2*F64)-4      17.0 ± 0%       3.0 ± 0%  -82.35%  (p=0.000 n=20+20)
FormulaFunc/42.0-4                 6.00 ± 0%      2.00 ± 0%  -66.67%  (p=0.000 n=20+20)
FormulaFunc/F64-4                  9.00 ± 0%      2.00 ± 0%  -77.78%  (p=0.000 n=20+20)
FormulaFunc/2*F64-4                10.0 ± 0%       2.0 ± 0%  -80.00%  (p=0.000 n=20+20)
FormulaFunc/math.Abs(2*F64)-4      15.0 ± 0%       2.0 ± 0%  -86.67%  (p=0.000 n=20+20)
```