@sbinet 2020-05-03
@rmadar. this should help a bit w/ your performance issues.

next, I'll change the API of `Formula` to take a pointer, as suggested in #647